requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118
torch==2.0.1+cu118
torchvision==0.15.2+cu118

accelerate==0.21.0
datasets==2.14.6
diffusers==0.23.0
einops==0.8.0
huggingface-hub==0.25.2
imageio==2.36.0
lpips==0.1.4
numpy==1.24.1
opencv-python==4.10.0.84
pillow==11.0.0
pyyaml==6.0.2
scikit-image==0.24.0
timm==1.0.11
tqdm==4.66.5
transformers==4.36.0
xformers==0.0.20
tensorboard==2.19.0